Updating Build changes

+++ src/com/example/calculator/Calculator.java
@@ -1,7 +1,7 @@
 package com.example.calculator;
 
 public class Calculator {
-	float add(float a, float b) {
+	int add(int a, int b) {
 		return a + b;
 	}
